+++ src/ZeppelinForms/Forms/Controls/ProgressBar.cs
@@ -42,7 +42,7 @@ public class ProgressBar : UnitControl, IBorderedElement
     }
 
     public Color FilledTextColor { get; set; } = Colors.White;
-    public Color FillColor { get; set; } = LightThemeColors.ButtonFill;
+    public Color FillColor { get; set; } = new Color(255, 0x0D, 0x6E, 0xFD);
     public Color TrackColor { get; set; } = new Color(255, 230, 230, 230);
     public Color TextColor { get; set; } = Colors.Black;