+++ src/ZeppelinForms/Forms/Controls/RadioButton.cs
@@ -24,7 +24,7 @@ public class RadioButton : UnitControl, ITextElement, IInputElement
     public Color TextColor { get; set; } = Colors.Black;
 
     public Color CircleBorderColor { get; set; } = Colors.Black;
-    public Color CheckColor { get; set; } = LightThemeColors.ButtonFill;
+    public Color CheckColor { get; set; } = new Color(255, 0x0D, 0x6E, 0xFD);
 
     // IInputElement
     public bool IsFocused { get; set; }