+++ src/ZeppelinForms/Forms/Controls/TabControl.cs
@@ -49,7 +49,7 @@ public class TabControl : PanelControl, IInputElement, IBorderedElement
     public Color SelectedHeaderColor { get; set; } = Colors.White;
     public Color TextColor { get; set; } = Colors.Black;
     public Color DisabledTextColor { get; set; } = new Color(255, 165, 165, 165);
-    public Color AccentColor { get; set; } = LightThemeColors.ButtonFill;
+    public Color AccentColor { get; set; } = new Color(255, 0x0D, 0x6E, 0xFD);
 
     public Color BorderColor { get; set; } = new Color(255, 205, 205, 205);
     public float BorderWidth { get; set; } = 1f;